Hi all,
In Ex2, Q2, please note that unknown words are words that appear in the test set but did not appear in the training set.

Omri
-----------------
As I understand Omirs remark no words in the training set are "unknown" words, only from the test set.
So I guess we'll need to remove that elif:
elif w in words_likely_tags and t == UNKNOWN_TAG: